feat(lib): Moves claims and registry code into wash lib
Sorry for the big PR here, but due to a bunch of codependent code,
I had to move a bunch of stuff at once. There are two main threads
to this PR. First, I noticed that the claims code is all CLI specific,
but it is likely that anyone building a CLI will not want to rewrite that
again. If you are doing this purely in code, you can just use the
wascap library. To make this work, I started added the CLI specific stuff
to the `cli` module of wash lib. There will probably be other things we
add to it as we finish this refactor

Second, this moves the reusable registry bits into its own module, which
is super handy even for those not doing a CLI as it avoids direct
interaction with the lower level OCI crates
