+++ crates/wash-lib/Cargo.toml
@@ -17,7 +17,7 @@ maintenance = {status = "actively-developed"}
 default = ["start", "parser"]
 start = ["semver"]
 parser = ["config", "semver", "serde", "serde_json"]
-cli = ["clap"]
+cli = ["clap", "term-table"]
 
 [dependencies]
 anyhow = "1.0.58"
@@ -29,6 +29,8 @@ dirs = "4.0"
 futures = "0.3"
 log = "0.4"
 nkeys = "0.2.0"
+oci-distribution = { version = "0.9.1", default-features = false, features = ["rustls-tls"] }
+provider-archive = "0.6.0"
 reqwest = { version = "0.11", default-features = false, features = ["json", "rustls-tls", "stream"] }
 semver = { version = "1.0.12", features = ["serde"], optional = true }
 serde = { version = "1", features = ["derive"], optional = true }
@@ -36,9 +38,11 @@ serde_json = { version = "1.0.82", optional = true }
 serde_with = "2.0.0"
 test-case = "2.2.1"
 thiserror = "1.0"
+term-table = { version = "1.3.1", optional = true }
 tokio = { version = "1", default-features = false, features = ["process"] }
 tokio-stream = "0.1"
 tokio-tar = "0.3"
+wascap = "0.8.0"
 weld-codegen = "0.5.0"
 
 [dev-dependencies]
