+++ src/appearance/spinner.rs
@@ -1,6 +1,6 @@
-use crate::util::OutputKind;
 use anyhow::Result;
 use indicatif::{ProgressBar, ProgressStyle};
+use wash_lib::cli::OutputKind;
 
 // For more spinners check out the cli-spinners project:
 // https://github.com/sindresorhus/cli-spinners/blob/master/spinners.json
